Fix DriverIT problem. skip AcceptanceTest from license check
