+++ driver/src/test/scala/com/stratio/crossdata/driver/DriverIT.scala
@@ -17,7 +17,6 @@
 package com.stratio.crossdata.driver
 
 import java.nio.file.Paths
-import java.util.concurrent.TimeoutException
 
 import akka.util.Timeout
 import com.stratio.crossdata.common.SQLCommand
@@ -26,7 +25,6 @@ import org.apache.spark.sql.AnalysisException
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 
-import scala.concurrent.Await
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
@@ -57,8 +55,8 @@ class DriverIT extends EndToEndTest {
     val sqlCommand = SQLCommand("SELECT * FROM jsonTable")
     val result = driver.syncQuery(sqlCommand)
     result shouldBe an[SuccessfulQueryResult]
-    result.queryId should be (sqlCommand.queryId)
-    result.hasError should be (false)
+    result.queryId should be(sqlCommand.queryId)
+    result.hasError should be(false)
     val rows = result.resultSet
     rows should have length 2
     rows(0) should have length 2
@@ -70,22 +68,5 @@ class DriverIT extends EndToEndTest {
     */
   }
 
-  "Crossdata driver" should "fail with a timeout when there is no server" in {
-    val driver = Driver()
-    val sqlCommand = SQLCommand("select * from any")
-    val result = driver.syncQuery(sqlCommand, Timeout(1 seconds), 1)
-    result.hasError should be (true)
-    a [RuntimeException] should be thrownBy result.resultSet
-    result.queryId should be (sqlCommand.queryId)
-    result shouldBe an [ErrorResult]
-    result.asInstanceOf[ErrorResult].message should include regex "(?i)timeout was exceed"
 
-  }
-
-
-  it should "return a future with a timeout when there is no server" in {
-    val driver = Driver()
-    val future = driver.asyncQuery(SQLCommand("select * from any"), Timeout(1 seconds), 1)
-    a [TimeoutException] should be thrownBy Await.result(future, 2 seconds)
-  }
 }
